Add update user functionality

+++ frontend/src/Pages/Admin/UsersListPage.js
@@ -11,6 +11,14 @@ import {
   Typography,
   TableHead,
   CircularProgress,
+  Modal,
+  Fade,
+  Backdrop,
+  TextField,
+  Button,
+  FormControlLabel,
+  Checkbox,
+  Alert,
 } from "@mui/material";
 import { Box } from "@mui/system";
 import { listUsers } from "redux/actions/userActions";
@@ -20,14 +28,41 @@ import EditIcon from "@mui/icons-material/Edit";
 import CheckIcon from "@mui/icons-material/Check";
 import CloseIcon from "@mui/icons-material/Close";
 import { deleteUser } from "redux/actions/userActions";
+import { getUserDetails } from "redux/actions/userActions";
+import { updateUser } from "redux/actions/userActions";
+
+const style = {
+  position: "absolute",
+  top: "50%",
+  left: "50%",
+  transform: "translate(-50%, -50%)",
+  width: 400,
+  bgcolor: "background.paper",
+  boxShadow: 24,
+  p: 4,
+  borderRadius: 1,
+};
 
 const UserListPage = () => {
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(5);
+  const [isModal, setIsModal] = useState(false);
+  const [name, setName] = useState("");
+  const [email, setEmail] = useState("");
+  const [isAdmin, setIsAdmin] = useState(false);
 
   const dispatch = useDispatch();
   const { users, loading } = useSelector((state) => state.userList);
+  const { user, loading: userLoading } = useSelector(
+    (state) => state.userDetails
+  );
   const { success: deleteSuccess } = useSelector((state) => state.userDelete);
+
+  const {
+    loading: updateLoading,
+    error: updateError,
+    success: updateSuccess,
+  } = useSelector((state) => state.userUpdate);
 
   //Change page
   const handleChangePage = (event, newPage) => {
@@ -41,12 +76,41 @@ const UserListPage = () => {
 
   //Delete user
   const handleDelete = (id) => {
-    dispatch(deleteUser(id));
+    if (window.confirm("Are you sure")) {
+      dispatch(deleteUser(id));
+    }
   };
 
+  //Edit user
+  const handleEdit = (id) => {
+    setIsModal(true);
+    dispatch(getUserDetails(id));
+  };
+
+  //Edit submit
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+
+    dispatch(updateUser({ _id: user._id, name, email, isAdmin }));
+
+    if (updateSuccess) {
+      setIsModal(false);
+    }
+  };
+
+  // render/re render user list
   useEffect(() => {
     dispatch(listUsers());
-  }, [dispatch, deleteSuccess]);
+  }, [dispatch, deleteSuccess, updateSuccess]);
+
+  useEffect(() => {
+    if (user) {
+      setName(user.name);
+      setEmail(user.email);
+      setIsAdmin(user.isAdmin);
+    }
+  }, [dispatch, user]);
 
   if (loading)
     return (
@@ -63,66 +127,134 @@ const UserListPage = () => {
     );
 
   return (
-    <Box sx={{ width: "100%" }}>
-      <Typography variant="h4" component="h1" sx={{ marginY: 5 }}>
-        User List
-      </Typography>
+    <>
+      <Modal
+        open={isModal}
+        onClose={() => setIsModal(false)}
+        closeAfterTransition
+        BackdropComponent={Backdrop}
+        BackdropProps={{
+          timeout: 500,
+        }}
+      >
+        <Fade in={isModal}>
+          <Box sx={style}>
+            <Typography variant="h4" component="h2">
+              Edit User
+            </Typography>
 
-      <Paper sx={{ width: "100%", mb: 2 }}>
-        <TableContainer>
-          <Table size="medium">
-            <TableHead>
-              <TableRow>
-                <TableCell></TableCell>
-                {["ID", "Email", "Name", "Admin"].map((headCell) => (
-                  <TableCell key={headCell}>{headCell}</TableCell>
-                ))}
-                <TableCell></TableCell>
-              </TableRow>
-            </TableHead>
+            {updateError && <Alert severity="error">{updateError}</Alert>}
 
-            <TableBody>
-              {users
-                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                .map((row, index) => {
-                  const labelId = `enhanced-table-checkbox-${index}`;
+            {userLoading ? (
+              "loading..."
+            ) : (
+              <Box
+                component="form"
+                mt={3}
+                sx={{
+                  "& > :not(style)": { my: 1, width: "100%" },
+                }}
+                onSubmit={handleSubmit}
+              >
+                <TextField
+                  label="Name"
+                  variant="outlined"
+                  value={name || ""}
+                  onChange={(e) => setName(e.target.value)}
+                />
 
-                  return (
-                    <TableRow hover key={row._id}>
-                      <TableCell></TableCell>
-                      <TableCell component="th" id={labelId} scope="row">
-                        {row._id}
-                      </TableCell>
-                      <TableCell>{row.email}</TableCell>
-                      <TableCell>{row.name}</TableCell>
-                      <TableCell>
-                        {row.isAdmin ? <CheckIcon /> : <CloseIcon />}
-                      </TableCell>
-                      <TableCell>
-                        <IconButton>
-                          <EditIcon />
-                        </IconButton>
-                        <IconButton onClick={() => handleDelete(row._id)}>
-                          <DeleteIcon />
-                        </IconButton>
-                      </TableCell>
-                    </TableRow>
-                  );
-                })}
-            </TableBody>
-          </Table>
-        </TableContainer>
-        <TablePagination
-          rowsPerPageOptions={[5, 10, 25]}
-          component="div"
-          count={users.length}
-          rowsPerPage={rowsPerPage}
-          page={page}
-          onPageChange={handleChangePage}
-          onRowsPerPageChange={handleChangeRowsPerPage}
-        />
-      </Paper>
-    </Box>
+                <TextField
+                  label="Email"
+                  variant="outlined"
+                  value={email || ""}
+                  onChange={(e) => setEmail(e.target.email)}
+                />
+
+                <FormControlLabel
+                  control={
+                    <Checkbox
+                      checked={isAdmin || false}
+                      onChange={(e) => setIsAdmin(e.target.checked)}
+                    />
+                  }
+                  label="Is Admin"
+                />
+
+                <Button
+                  variant="contained"
+                  type="submit"
+                  size="large"
+                  sx={{ height: "45px" }}
+                >
+                  {updateLoading ? "loading..." : "UPDATE"}
+                </Button>
+              </Box>
+            )}
+          </Box>
+        </Fade>
+      </Modal>
+
+      <Box sx={{ width: "100%" }}>
+        <Typography variant="h4" component="h1" sx={{ marginY: 5 }}>
+          User List
+        </Typography>
+
+        <Paper sx={{ width: "100%", mb: 2 }}>
+          <TableContainer>
+            <Table size="medium">
+              <TableHead>
+                <TableRow>
+                  <TableCell></TableCell>
+                  {["ID", "Email", "Name", "Admin"].map((headCell) => (
+                    <TableCell key={headCell}>{headCell}</TableCell>
+                  ))}
+                  <TableCell></TableCell>
+                </TableRow>
+              </TableHead>
+
+              <TableBody>
+                {users
+                  .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+                  .map((row, index) => {
+                    const labelId = `enhanced-table-checkbox-${index}`;
+
+                    return (
+                      <TableRow hover key={row._id}>
+                        <TableCell></TableCell>
+                        <TableCell component="th" id={labelId} scope="row">
+                          {row._id}
+                        </TableCell>
+                        <TableCell>{row.email}</TableCell>
+                        <TableCell>{row.name}</TableCell>
+                        <TableCell>
+                          {row.isAdmin ? <CheckIcon /> : <CloseIcon />}
+                        </TableCell>
+                        <TableCell>
+                          <IconButton onClick={() => handleEdit(row._id)}>
+                            <EditIcon />
+                          </IconButton>
+                          <IconButton onClick={() => handleDelete(row._id)}>
+                            <DeleteIcon />
+                          </IconButton>
+                        </TableCell>
+                      </TableRow>
+                    );
+                  })}
+              </TableBody>
+            </Table>
+          </TableContainer>
+          <TablePagination
+            rowsPerPageOptions={[5, 10, 25]}
+            component="div"
+            count={users.length}
+            rowsPerPage={rowsPerPage}
+            page={page}
+            onPageChange={handleChangePage}
+            onRowsPerPageChange={handleChangeRowsPerPage}
+          />
+        </Paper>
+      </Box>
+    </>
   );
 };
